fix transaction item expansion and remove balance calculation
